+++ RedPandaIDE/systemconsts.h
@@ -3,7 +3,7 @@
 
 #include <QStringList>
 
-#define DEVCPP_VERSION "beta.0.8.5"
+#define DEVCPP_VERSION "beta.0.8.6"
 
 #define APP_SETTSINGS_FILENAME "redpandacpp.ini"
 #ifdef Q_OS_WIN